# Conflicts:
#	TotalCrossVM/src/init/noras_ids/noras.inc

+++ TotalCrossVM/src/init/darwin/NSString+MD5Addition.m
@@ -19,7 +19,7 @@
     const char *value = [self UTF8String];
     
     unsigned char outputBuffer[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(value, strlen(value), outputBuffer);
+    CC_MD5(value, (int)strlen(value), outputBuffer);
     
     NSMutableString *outputString = [[NSMutableString alloc] initWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(NSInteger count = 0; count < CC_MD5_DIGEST_LENGTH; count++){
